+++ kbe/src/server/cellapp/clients_remote_entity_method.cpp
@@ -101,7 +101,7 @@ PyObject* ClientsRemoteEntityMethod::callmethod(PyObject* args, PyObject* kwds)
 			pEntity->clientMailbox()->newMail((*pBundle));
 
 			if(mstream->wpos() > 0)
-				(*pBundle).append(mstream->data(), mstream->wpos());
+				(*pBundle).append(mstream->data(), (int)mstream->wpos());
 
 			if(Network::g_trace_packet > 0)
 			{
@@ -154,7 +154,9 @@ PyObject* ClientsRemoteEntityMethod::callmethod(PyObject* args, PyObject* kwds)
 			if(pChannel == NULL)
 				continue;
 
-			if(!pAoiEntity->pWitness()->entityInAOI(pEntity->id()))
+			// 这个可能性是存在的，例如数据来源于createWitnessFromStream()
+			// 又如自己的entity还未在目标客户端上创建
+			if (!pAoiEntity->pWitness()->entityInAOI(pEntity->id()))
 				continue;
 
 			Network::Bundle* pSendBundle = Network::Bundle::ObjPool().createObject();
@@ -164,7 +166,7 @@ PyObject* ClientsRemoteEntityMethod::callmethod(PyObject* args, PyObject* kwds)
 					ClientInterface::onRemoteMethodCallOptimized, pEntity->id());
 
 			if(mstream->wpos() > 0)
-				(*pForwardBundle).append(mstream->data(), mstream->wpos());
+				(*pForwardBundle).append(mstream->data(), (int)mstream->wpos());
 
 			if(Network::g_trace_packet > 0)
 			{
